% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.DALSM.R
\name{AIC.DALSM}
\alias{AIC.DALSM}
\title{Akaike Information Criterion (AIC) of a DALSM object.}
\usage{
\method{AIC}{DALSM}(object, ..., k=2)
}
\arguments{
\item{object}{A \code{\link{DALSM.object}}.}

\item{...}{Other optional DALSM objects.}

\item{k}{The penalty per parameter to be used. (Default: k=2 for the classical AIC).}
}
\value{
The AIC of the fitted DALSM model in \code{x}.
}
\description{
Akaike Information Criterion (AIC) for the fitted DALSM model in a \code{DALSM.object}.
}
\details{
Akaike information criterion for the fitted model in a DALSM object, with a penalty calculated using the total effective degrees of freedom, -2log(L) + 2*ED.tot, smaller values being preferred during model selection.
}
\examples{
require(DALSM)
data(DALSM_IncomeData)
resp = DALSM_IncomeData[,1:2]
fit = DALSM(y=resp,
            formula1 = ~twoincomes+s(age)+s(eduyrs),
            formula2 = ~twoincomes+s(age)+s(eduyrs),
            data = DALSM_IncomeData)
AIC(fit)

}
\references{
Time-varying exogenous covariates with frequently changing values in double additive cure survival model: an application to fertility.
\emph{Journal of the Royal Statistical Society, Series A}. <doi:10.1093/jrsssa/qnaf035>
}
\seealso{
\code{\link{DALSM}}, \code{\link{DALSM.object}}, \code{\link{BIC.DALSM}}, \code{\link{logEvid}}
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
