% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.DALSM.R
\name{BIC.DALSM}
\alias{BIC.DALSM}
\title{Bayesian Information Criterion (BIC) of a DALSM object.}
\usage{
\method{BIC}{DALSM}(object, ...)
}
\arguments{
\item{object}{A \code{\link{DALSM.object}}.}

\item{...}{Other optional DALSM objects.}
}
\value{
The BIC of the fitted DALSM model in \code{x}.
}
\description{
Bayesian Information Criterion (BIC) for the fitted DALSM model in a \code{DALSM.object}.
}
\details{
Bayesian (Schwarz) information for the fitted model in a DALSM object, with a penalty calculated using the total effective degrees of freedom and the total number of observed or interval-censored data, -2log(L) + log(n.uncensored)*ED.tot, smaller values being preferred during model selection.
}
\examples{
require(DALSM)
data(DALSM_IncomeData)
resp = DALSM_IncomeData[,1:2]
fit = DALSM(y=resp,
            formula1 = ~twoincomes+s(age)+s(eduyrs),
            formula2 = ~twoincomes+s(age)+s(eduyrs),
            data = DALSM_IncomeData)
BIC(fit)

}
\references{
Time-varying exogenous covariates with frequently changing values in double additive cure survival model: an application to fertility.
\emph{Journal of the Royal Statistical Society, Series A}. <doi:10.1093/jrsssa/qnaf035>
}
\seealso{
\code{\link{DALSM}}, \code{\link{DALSM.object}}, \code{\link{BIC.DALSM}}, \code{\link{logEvid}}
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
