% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverageAdditiveTerms.R
\name{CoverageAdditiveTerms}
\alias{CoverageAdditiveTerms}
\title{Fitted \code{DALSM} additive terms and effective coverage of their credible intervals}
\usage{
CoverageAdditiveTerms(obj.fit, nx=101,
                      true.loc=NULL, true.disp=NULL, ci.level=NULL,
                      silent=FALSE)
}
\arguments{
\item{obj.fit}{an object of class \code{\link{DALSM.object}}}

\item{nx}{(optional) grid size of covariate values where the additive terms are calculated (default: 101)}

\item{true.loc}{(optional) list of functions containing the 'true' additive terms for location}

\item{true.disp}{(optional) list of functions containing the 'true' additive terms for dispersion}

\item{ci.level}{(optional) level of credible intervals}

\item{silent}{logical indicating whether the computed corverages should be printed out (default: FALSE)}
}
\value{
It returns an invisible list containing:
\itemize{
\item{\code{J1} : \verb{ }}{number of additive terms in the location sub-model}
\item{\code{x.loc} : \verb{ }}{a \code{nx} by \code{J1} matrix containing a regular grid of \code{nx} covariate values where the corresponding additive term in location is evaluated}
\item{\code{f.loc} : \verb{ }}{a \code{nx} by \code{J2} matrix containing the \code{J2} fitted location additive terms evaluated at \code{x.loc}}
\item{\code{se.loc} : \verb{ }}{a \code{nx} by \code{J2} matrix containing the the pointwise standard errors of the fitted location additive terms evaluated at \code{x.loc}}
\item{\code{coverage.loc} : \verb{ }}{a vector of length length \code{J1} giving the average effective coverage of pointwise credible intervals for each of the additive terms for location}
\item{\code{J2} : \verb{ }}{number of additive terms in the dispersion sub-model}
\item{\code{x.disp} : \verb{ }}{a \code{nx} by \code{J2} matrix containing a regular grid of \code{nx} covariate values where the corresponding additive term in dispersion is evaluated}
\item{\code{f.disp} : \verb{ }}{a \code{nx} by \code{J2} matrix containing the \code{J2} fitted dispersion additive terms evaluated at \code{x.disp}}
\item{\code{se.disp} : \verb{ }}{a \code{nx} by \code{J2} matrix containing the pointwise standard errors of the fitted dispersion additive terms evaluated at \code{x.disp}}
\item{\code{coverage.disp} : \verb{ }}{a vector of length length \code{J2} giving the average effective coverage of pointwise credible intervals for each of the additive terms for dispersion}
}
}
\description{
Compute the fitted additive terms of a \code{DALSM} model on a regular grid and report the mean effective coverage of pointwise credible intervals
for these additive terms with respect to given reference functions (such as the 'true' additive terms used in a simulation study to generate the data)
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}#'
}
