% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dens1d.R
\name{Dens1d}
\alias{Dens1d}
\title{Creates an object for density estimation from right- or interval-censored data using \code{\link{densityGivenMeanVariance}}}
\usage{
Dens1d(
  y,
  event = NULL,
  ymin = NULL,
  ymax = NULL,
  K = 25,
  equid.knots = TRUE,
  pen.order = 2,
  nbins = 501
)
}
\arguments{
\item{y}{a n-vector (if no interval-censored data) or a nx2 matrix (left and right limits of the interval-censored variable ; identical columns if right-censored)}

\item{event}{a n-vector of observation indicators (0: right-censored ; 1: exactly observed or interval-censored)}

\item{ymin}{left limit of the variable support}

\item{ymax}{right limit of the variable support}

\item{K}{number of B-splines in the basis to approximate the log-hazard}

\item{equid.knots}{logical indicating if equidistants knots are desired}

\item{pen.order}{penalty order when equidistant knots (otherwise: computed from the second derivative)}

\item{nbins}{number of small bins for quadrature and approximations}
}
\value{
A \code{Dens1d} object, i.e. a list with summary measures and precomputed components required for density estimation using \code{\link{densityGivenMeanVariance}}
}
\description{
Creates an object for density estimation from right- or interval-censored data using \code{\link{densityGivenMeanVariance}}
}
