% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityGivenMeanVariance_object.R
\name{densityGivenMeanVariance.object}
\alias{densityGivenMeanVariance.object}
\title{Object resulting from the fit of a constrained density on potentially right- or interval-censored data.}
\value{

}
\description{
An object returned by the \code{\link{densityGivenMeanVariance}} function: this is a list
with various components related to the fit of a constrained density on potentially right- or interval-censored data.
}
\seealso{
\code{\link{DALSM}}, \code{\link{print.DALSM}}, \code{\link{plot.DALSM}}
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
