% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logDet_fun.R
\name{logDet.fun}
\alias{logDet.fun}
\title{Log-determinant of a positive-definite matrix}
\usage{
logDet.fun(x)
}
\arguments{
\item{x}{positive definite matrix.}
}
\value{
log of det(x).
}
\description{
Log-determinant of a positive-definite matrix
}
\examples{
A = matrix(1:4,ncol=2)
logDet.fun(A)

}
\references{
Lambert, P. (2021). Fast Bayesian inference using Laplace approximations
in nonparametric double additive location-scale models with right- and
interval-censored data.
\emph{Computational Statistics and Data Analysis}, 161: 107250.
<doi:10.1016/j.csda.2021.107250>
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
