% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logEvid.R
\name{logEvid.DALSM}
\alias{logEvid.DALSM}
\title{Log-evidence of a DALSM object.}
\usage{
\method{logEvid}{DALSM}(object, ...)
}
\arguments{
\item{object}{A \code{\link{DALSM.object}}.}

\item{...}{Optionally more DALSM objects.}
}
\value{
The log-evidence of the DALSM model in \code{object}.
}
\description{
The log-evidence of the fitted DALSM model in a \code{DALSM.object}.
}
\details{
Provides the log-evidence (or log-marginal likelihood) of the fitted DALSM model in a given \code{\link{DALSM.object}}, where the evidence is the marginal posterior of the penalty parameters at their selected values.
}
\examples{
require(DALSM)
data(DALSM_IncomeData)
resp = DALSM_IncomeData[,1:2]
fit = DALSM(y=resp,
            formula1 = ~twoincomes+s(age)+s(eduyrs),
            formula2 = ~twoincomes+s(age)+s(eduyrs),
            data = DALSM_IncomeData)
logEvid(fit)

}
\references{
Time-varying exogenous covariates with frequently changing values in double additive cure survival model: an application to fertility.
\emph{Journal of the Royal Statistical Society, Series A}. <doi:10.1093/jrsssa/qnaf035>
}
\seealso{
\code{\link{DALSM}}, \code{\link{DALSM.object}}, \code{\link{AIC.DALSM}}, \code{\link{BIC.DALSM}}
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
