% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pcal_fun.R
\name{Pcal.fun}
\alias{Pcal.fun}
\title{Compute the penalty matrix associated to a vector containing fixed (non-penalized) parameters and equal-size sub-vectors of penalized spline parameters}
\usage{
Pcal.fun(nfixed, lambda, Pd.x)
}
\arguments{
\item{nfixed}{the number of fixed (i.e. non-penalized) parameters}

\item{lambda}{a vector of \code{p} penalty parameters where each component is associated to a sub-vector of spline parameters of length \code{J}}

\item{Pd.x}{a penalty matrix of size \code{J} associated to a given sub-vector of spline parameters}
}
\value{
A block diagonal penalty matrix of size \code{(nfixed+pJ)} given by Blockdiag(diag(0,\code{nfixed}), diag(\code{lambda}).kron.\code{Pd.x})
}
\description{
Compute the penalty matrix associated to a vector containing fixed (non-penalized) parameters and equal-size sub-vectors of penalized spline parameters
}
