% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.DALSM2.R
\name{plot.DALSM}
\alias{plot.DALSM}
\title{Plot visual information on a \code{DALSM.object}}
\usage{
\method{plot}{DALSM}(x, ngrid=300, ci.level=.95, pages=0, select=NULL,
       fill=TRUE, pointwise=TRUE, mar=c(4,5,1,1),
       xlim0=NULL, ylim0=NULL, xlim1=NULL, ylim1=NULL, xlim2=NULL, ylim2=NULL,
       equal.ylims=TRUE, ...)
}
\arguments{
\item{x}{a \code{\link{DALSM.object}}.}

\item{ngrid}{(optional) number of points to make the plots for the fitted additive terms (default: 300).}

\item{ci.level}{(optional) nominal level for the plotted pointwise credible intervals (default: .95).}

\item{pages}{The number of pages over which to spread the output. For example, if pages=1 then all terms will be plotted on one page with the layout performed automatically. Set to 0 to have the routine leave all graphics settings as they are (default: 0).}

\item{select}{Allows the plot for a single model term to be selected for printing. e.g. if you just want the plot for the second smooth term set select=2. The plot of the standardized error density \eqn{f_0(z)} is provided when select=0 (default: NULL).}

\item{fill}{Logical indicating whether credible regions should be greyed out. (Default: TRUE).}

\item{pointwise}{Logical indicating whether only pointwise credible intervals should be plotted for additive terms. When FALSE, simultaneous credible regions are also provided. (Default: TRUE).}

\item{mar}{A numerical vector of the form c(bottom, left, top, right) which gives the number of lines of margin to be specified on the four sides of the plot. (Default: c(4,5,1,1)).}

\item{xlim0}{Vector of length 2 specifying x-axis limits when plotting the estimated error density function \eqn{f_0(z)}.}

\item{ylim0}{Vector of length 2 specifying y-axis limits when plotting the estimated error density function \eqn{f_0(z)}.}

\item{xlim1}{Vector of length 2 specifying (common) x-axis limits when plotting the fitted additive term(s) entering the conditional mean submodel. (Default: NULL).}

\item{ylim1}{Vector of length 2 specifying (common) y-axis limits when plotting the fitted additive term(s) entering the conditional mean submodel. (Default: NULL).}

\item{xlim2}{Vector of length 2 specifying (common) x-axis limits when plotting the fitted additive term(s) in the log of the conditional standard deviation submodel. (Default: NULL).}

\item{ylim2}{Vector of length 2 specifying (common) y-axis limits when plotting the fitted additive term(s) entering the conditional standard deviation submodel. (Default: NULL).}

\item{equal.ylims}{logical indicating if the same y-limits must be used when plotting the fitted additive terms from the same submodel. It can be overriden by non-NULL values for \code{ylim1} or \code{ylim2}. (Default: TRUE).}

\item{...}{additional generic plotting arguments.}
}
\value{
In addition to the plots, an invisible list containing the following is returned:
\itemize{
\item \code{J1} : number of additive terms in the location sub-model.
\item \code{x.loc} : a \code{ngrid} by \code{J1} matrix containing a regular grid of \code{ngrid} covariate values where the corresponding additive term in location is evaluated.
\item \code{f.loc} : a \code{ngrid} by \code{J1} matrix containing the \code{J1} fitted location additive terms evaluated at \code{x.loc}.
\item \code{se.loc} : a \code{ngrid} by \code{J1} matrix containing the the pointwise standard errors of the fitted location additive terms evaluated at \code{x.loc}.
\item \code{J2} : number of additive terms in the dispersion sub-model.
\item \code{x.disp} : a \code{ngrid} by \code{J2} matrix containing a regular grid of \code{ngrid} covariate values where the corresponding additive term in dispersion is evaluated.
\item \code{f.disp} : a \code{ngrid} by \code{J2} matrix containing the \code{J2} fitted dispersion additive terms evaluated at \code{x.disp}.
\item \code{se.disp} : a \code{ngrid} by \code{J2} matrix containing the pointwise standard errors of the fitted dispersion additive terms evaluated at \code{x.disp}.
}
}
\description{
Visualize the estimated additive terms and error density corresponding to a Double Additive Location-Scale Model (DALSM) object.
}
\details{
Plot the fitted additive terms and the estimated standardized error density contained in the \code{\link{DALSM.object}} \code{x}.
}
\examples{
require(DALSM)
data(DALSM_IncomeData)
resp = DALSM_IncomeData[,1:2]
fit = DALSM(y=resp,
            formula1 = ~twoincomes+s(age)+s(eduyrs),
            formula2 = ~twoincomes+s(age)+s(eduyrs),
            data = DALSM_IncomeData)
plot(fit, select=0) ## Estimated error density
plot(fit, select=1:4, pages=1) ## Estimated additive terms

}
\references{
Lambert, P. (2021). Fast Bayesian inference using Laplace approximations
in nonparametric double additive location-scale models with right- and
interval-censored data.
\emph{Computational Statistics and Data Analysis}, 161: 107250.
<doi:10.1016/j.csda.2021.107250>
}
\seealso{
\code{\link{DALSM}}, \code{\link{DALSM.object}}, \code{\link{print.DALSM}}
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
