% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.DALSM.R
\name{print.DALSM}
\alias{print.DALSM}
\title{Print summary information on a fitted Double Additive Location-Scale Model (DALSM) from a DALSM object}
\usage{
\method{print}{DALSM}(x,...)
}
\arguments{
\item{x}{an object of class \code{\link{DALSM.object}}}

\item{...}{additional generic printing arguments}
}
\description{
Print summary information on a fitted Double Additive Location-Scale Model (DALSM) from a DALSM object
}
\details{
Provides summary measures on the estimation of the regression parameters and additive terms
in the location and dispersion submodels corresponding to a \code{\link{DALSM.object}} generated by \code{\link{DALSM}}
}
\seealso{
\code{\link{DALSM}}, \code{\link{DALSM.object}}, \code{\link{plot.DALSM}}
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
