% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.densityGivenMeanVariance.R
\name{print.densityGivenMeanVariance}
\alias{print.densityGivenMeanVariance}
\title{Print summary information on the error density of a fitted DALSM regression model}
\usage{
\method{print}{densityGivenMeanVariance}(x,...)
}
\arguments{
\item{x}{a <densityGivenMeanVariance.object> (typically the $derr element of a DALSM.object list)}

\item{...}{Optional additional print parameters}
}
\value{
No returned value (just printed summary)
}
\description{
Print summary information on the error density of a fitted DALSM regression model
}
\examples{
require(DALSM)
data(DALSM_IncomeData)
resp = DALSM_IncomeData[,1:2]
fit = DALSM(y=resp,
            formula1 = ~twoincomes+s(age)+s(eduyrs),
            formula2 = ~twoincomes+s(age)+s(eduyrs),
            data = DALSM_IncomeData)
plot(fit$derr)  ## Plot the estimated error density
print(fit$derr) ## ... and provide some descriptive elements on it

}
\references{
Lambert, P. (2021). Fast Bayesian inference using Laplace approximations
in nonparametric double additive location-scale models with right- and
interval-censored data.
\emph{Computational Statistics and Data Analysis}, 161: 107250.
\url{https://doi.org/10.1016/j.csda.2021.107250}
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
