% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.hist.R
\name{weighted.hist}
\alias{weighted.hist}
\title{Histogram from weighted data}
\usage{
weighted.hist(x, weight = NULL, breaks = "Sturges", freq = NULL,
probability = !freq, include.lowest = TRUE, right = TRUE, fuzz = 1e-07,
density = NULL, angle = 45, col = "lightgray", border = NULL,
main = paste("Histogram of", xname), xlim = range(breaks), ylim = NULL,
xlab = xname, ylab, axes = TRUE, plot = TRUE, labels = FALSE, nclass = NULL,
warn.unused = TRUE, ...)
}
\arguments{
\item{x}{a vector of values for which the histogram is desired.}

\item{weight}{optional vector of the same length as \code{x} defining the weights associated each each element in \code{x}. By default, it is \code{rep(1,length(x))}.}

\item{breaks}{one of:
\itemize{
   \item a vector giving the breakpoints between histogram cells,
   \item a single number giving the number of cells for the histogram,
   \item a character string naming an algorithm to compute the number of cells (see \sQuote{Details}),
   \item a function to compute the number of cells.
}}

\item{freq}{logical; if \code{TRUE}, the histogram graphic is a representation of frequencies, the \code{counts} component of the result; if \code{FALSE}, probability densities, component density, are plotted (so that the histogram has a total area of one). Defaults to \code{TRUE} \emph{if and only if} \code{breaks} are equidistant (and \code{probability} is not specified).}

\item{probability}{an \emph{alias} for \code{!freq}, for S compatibility.}

\item{include.lowest}{logical; if \code{TRUE}, an \code{x[i]} equal to the \code{breaks} value will be included in the first (or last, for \code{right = FALSE}) bar. This will be ignored (with a warning) unless \code{breaks} is a vector.}

\item{right}{logical; if \code{TRUE}, the histogram cells are right-closed (left open) intervals.}

\item{fuzz}{non-negative number, for the case when the data is \emph{pretty} and some observations \code{x[.]} are close but not exactly on a \code{break}. For counting fuzzy breaks proportional to \code{fuzz} are used. The default is occasionally suboptimal.}

\item{density}{the density of shading lines, in lines per inch. The default value of \code{NULL} means that no shading lines are drawn. Non-positive values of \code{density} also inhibit the drawing of shading lines.}

\item{angle}{the slope of shading lines, given as an angle in degrees (counter-clockwise).}

\item{col}{a colour to be used to fill the bars.}

\item{border}{the color of the border around the bars.  The default is to use the standard foreground color.}

\item{main}{main title.}

\item{xlim}{range of \code{x} for plotting.}

\item{ylim}{range of \code{y} values.}

\item{xlab}{label for the horizontal axis.}

\item{ylab}{label for the vertical axis.}

\item{axes}{logical.  If \code{TRUE} (default), axes are draw if the plot is drawn.}

\item{plot}{logical.  If \code{TRUE} (default), a histogram is plotted, otherwise a list of breaks and counts is returned.  In the latter case, a warning is used if (typically graphical) arguments are specified that only apply to the \code{plot=TRUE} case.}

\item{labels}{logical or character.  Additionally draw labels on top of bars, if not \code{FALSE}; see \code{plot.histogram} in the \code{graphics} package.}

\item{nclass}{numeric (integer).  For S(-PLUS) compatibility only, \code{nclass} is equivalent to \code{breaks} for a scalar or character argument.}

\item{warn.unused}{logical.  If \code{plot=FALSE} and \code{warn.unused=TRUE}, a warning will be issued when graphical parameters are passed to \code{hist.default()}.}

\item{...}{further arguments and graphical parameters passed to \code{plot.histogram} and thence to title and axis (if \code{plot=TRUE}).}
}
\value{
a list comparable to what \code{hist.default} produces and a plot the weighted histogram if \code{plot=TRUE}.
}
\description{
Computes a histogram from weighted data.
  If \code{plot=TRUE}, the resulting object of class \code{histogram} is
  plotted before it is returned.
}
\examples{
## Example 1
set.seed(123)
mround <- function(x,base) base*round(x/base) ## Rounding function
y = mround(rgamma(500,10,2), base=.5) ## Rounded data
tab = table(y) ; print(tab) ## Empirical distribution of the rounded data
## Histogram from the empirical distribution
res = weighted.hist(as.numeric(names(tab)), weight=c(tab), xlab="",main="")

## Example 2
## Generate categorized data
set.seed(123)
brks = c(0,3,5,7,9,15) ; ycat = cut(rgamma(500,10,2),breaks=brks)
tab = table(ycat) ## Empirical distribution of the categorized data
print(tab)
ymid = .5 * (brks[-6]+brks[-1]) ; w = c(tab) ## Midpoints and their weights
## Histogram from weighted data
res = weighted.hist(ymid,weight=w,breaks=brks,xlab="",main="")
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}. The code for this function is fully based on \code{hist.default} from the graphics package in R, with minor modifications to account for weights.
}
