% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityGivenMeanVariance.R
\name{densityGivenMeanVariance}
\alias{densityGivenMeanVariance}
\title{Constrained density estimation from potentially right- or interval-censored data, with given mean or variance.}
\usage{
densityGivenMeanVariance(
  obj.data,
  is.density = TRUE,
  Mean0 = NULL,
  Var0 = NULL,
  fixed.penalty = FALSE,
  method = c("evidence", "Schall"),
  fixed.phi = FALSE,
  phi.ref = NULL,
  phi0 = NULL,
  tau0 = exp(5),
  tau.min = 0.1,
  verbose = TRUE
)
}
\arguments{
\item{obj.data}{a list created from potentially right- or interval-censored data using \code{\link{Dens1d}}. It includes summary statistics, assumed density support, knots for the B-spline basis, etc.}

\item{is.density}{(optional) logical indicating whether the estimated density should integrate to 1.0 over the range of the knots in obj.data$knots (default: TRUE)}

\item{Mean0}{(optional) constrained value for the mean of the fitted density (defaut: NULL)}

\item{Var0}{(optional) constrained value for the variance of the fitted density (defaut: NULL)}

\item{fixed.penalty}{(optional) logical indicating whether the penalty parameter should be selected from the data (\code{fixed.penalty}=FALSE) or fixed (\code{fixed.penalty}=TRUE) at its initial value \code{tau0}}

\item{method}{method used for penaly selection: "evidence" or "Schall"}

\item{fixed.phi}{(optional) logical indicating whether the spline parameters are fixed (\code{fixed.phi}=TRUE) or estimated from the data (default: \code{fixed.phi}=FALSE)}

\item{phi.ref}{(optional) reference value for the spline parameters with respect to which deviations are penalized (default: zero vector)}

\item{phi0}{starting value for the spline parameters (default: spline parameters corresponding to a Student density with 5 DF)}

\item{tau0}{(optional) initial value for the penalty parameter (default: exp(5))}

\item{tau.min}{(optional) minimal value for the penalty parameter (default: .1)}

\item{verbose}{(optional) logical indicating whether estimation step details should be displayed (default: FALSE)}
}
\value{
a \code{\link{densityGivenMeanVariance.object}} containing the density estimation results
}
\description{
Constrained density estimation from potentially right- or interval-censored data, with given mean or variance.
}
