% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.DALSM.R
\name{plot.DALSM}
\alias{plot.DALSM}
\title{Plot a Double Additive Location-Scale Model (DALSM) object}
\usage{
\method{plot}{DALSM}(x,
       mfrow.loc=NULL, mfrow.disp=NULL,
       nx=101, plot.global=TRUE, true.loc=NULL,true.disp=NULL, ci.level=NULL,
       error.lim = NULL, plot.residuals=FALSE, true.derr=NULL, ...)
}
\arguments{
\item{x}{an object of class \code{\link{DALSM.object}}}

\item{mfrow.loc}{(optional) window layout to plot the additive terms for location}

\item{mfrow.disp}{(optional) window layout to plot the additive terms for dispersion}

\item{nx}{(optional) number of points to make the plots for the fitted additive terms (default: 101)}

\item{plot.global}{logical indicating if the same range of values must be used when plotting the fitted additive terms (default: TRUE)}

\item{true.loc}{(optional) list of functions to be superposed to the corresponding estimated additive terms in the location submodel (default: NULL)}

\item{true.disp}{(optional) list of functions to be superposed to the corresponding estimated additive terms in the dispersion submodel (default: NULL)}

\item{ci.level}{(optional) nominal level for the plotted pointwise or global credible intervals (default: x$ci.level)}

\item{error.lim}{(optional) plotting interval for the estimated standardized error density in the DALSM model (default: support of the fitted standardized error density)}

\item{plot.residuals}{logical requesting to add the (possibily censored) standardized residuals to the plot of the fitted standardized error density (default: FALSE)}

\item{true.derr}{(optional) density function to superpose to the estimated standardized error density (default: NULL)}

\item{...}{additional generic plotting arguments}
}
\value{
In addition to the plots, an invisible return of a list containing:
\itemize{
\item{\code{J1} : \verb{ }}{number of additive terms in the location sub-model}
\item{\code{x.loc} : \verb{ }}{a \code{nx} by \code{J1} matrix containing a regular grid of \code{nx} covariate values where the corresponding additive term in location is evaluated}
\item{\code{f.loc} : \verb{ }}{a \code{nx} by \code{J2} matrix containing the \code{J2} fitted location additive terms evaluated at \code{x.loc}}
\item{\code{se.loc} : \verb{ }}{a \code{nx} by \code{J2} matrix containing the the pointwise standard errors of the fitted location additive terms evaluated at \code{x.loc}}
\item{\code{J2} : \verb{ }}{number of additive terms in the dispersion sub-model}
\item{\code{x.disp} : \verb{ }}{a \code{nx} by \code{J2} matrix containing a regular grid of \code{nx} covariate values where the corresponding additive term in dispersion is evaluated}
\item{\code{f.disp} : \verb{ }}{a \code{nx} by \code{J2} matrix containing the \code{J2} fitted dispersion additive terms evaluated at \code{x.disp}}
\item{\code{se.disp} : \verb{ }}{a \code{nx} by \code{J2} matrix containing the pointwise standard errors of the fitted dispersion additive terms evaluated at \code{x.disp}}
}
}
\description{
Plot a Double Additive Location-Scale Model (DALSM) object
}
\details{
Plot the fitted additive terms and the estimated standardized error density contained in the \code{\link{DALSM.object}} \code{x}
}
\seealso{
\code{\link{DALSM}}, \code{\link{DALSM.object}}, \code{\link{print.DALSM}}
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
