% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.densityGivenMeanVariance.R
\name{plot.densityGivenMeanVariance}
\alias{plot.densityGivenMeanVariance}
\title{Plot the error density of a fitted DALSM regression model}
\usage{
\method{plot}{densityGivenMeanVariance}(
  x,
  xlim = range(fit$bins),
  breaks = NULL,
  histRC = FALSE,
  xlab = "",
  ylab = "Density",
  main = "",
  ...
)
}
\arguments{
\item{x}{a <densityGivenMeanVariance.object> (typically the $derr element of a DALSM.object list)}

\item{xlim}{interval of values where the density should be plotted}

\item{breaks}{(Optional) breaks for the histogram of the observed residuals}

\item{histRC}{Logical (Default: FALSE) indicating whether the histogram of the right-censored residuals should be highlighted}

\item{xlab}{Optional label for the x-axis}

\item{ylab}{Optional label for the y-axis}

\item{main}{Plot main title}

\item{...}{Optional additional plot parameters}
}
\value{
No returned value (just plots)
}
\description{
Plot the error density corresponding to the fit of a DALSM regression model. The histogram of the observed residuals can be superposed to it.
}
\examples{
require(DALSM)
data(DALSM_IncomeData)
resp = DALSM_IncomeData[,1:2]
fit = DALSM(y=resp,
            formula1 = ~twoincomes+s(age)+s(eduyrs),
            formula2 = ~twoincomes+s(age)+s(eduyrs),
            data = DALSM_IncomeData)
plot(fit$derr)  ## Plot the estimated error density
print(fit$derr) ## ... and provide some descriptive elements on it
}
\references{
Lambert, P. (2021). Fast Bayesian inference using Laplace approximations
in nonparametric double additive location-scale models with right- and
interval-censored data.
\emph{Computational Statistics and Data Analysis}, 161: 107250.
\url{https://doi.org/10.1016/j.csda.2021.107250}
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
