% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{s}
\alias{s}
\title{Internal function to specify smooth terms in DALSM formulas}
\usage{
s(x)
}
\arguments{
\item{x}{A variable for which an additive term is requested.}
}
\value{
Returns `x` unchanged.
}
\description{
This function serves as a placeholder for marking additive terms in formulas
passed to `DALSM()`. It does not perform any computation but allows users
to specify smooth terms explicitly.
}
\keyword{internal}
